backend/requirements.txt
django==4.2.*
psycopg2-binary
requests
djangorestframework
markdown
django-filter
pyyaml
